+++ src/sentry/workflow_engine/endpoints/validators/base/__init__.py
@@ -5,14 +5,9 @@ __all__ = [
     "BaseDataSourceValidator",
     "BaseDetectorTypeValidator",
     "DataSourceCreator",
-    "NumericComparisonConditionValidator",
 ]
 
 from .action import BaseActionValidator
-from .data_condition import (
-    BaseDataConditionGroupValidator,
-    BaseDataConditionValidator,
-    NumericComparisonConditionValidator,
-)
+from .data_condition import BaseDataConditionGroupValidator, BaseDataConditionValidator
 from .data_source import BaseDataSourceValidator, DataSourceCreator
 from .detector import BaseDetectorTypeValidator
